DELTA changes-to-allow-identifying-threads-in-log-file-output-20010404 AUTHOR reuter@rzg.mpg.de
"This is an addition for MR-AFS log files where the thread
name is shown. rx switches the general server threads to listeners and
vice versa. To have reasonable names the listener must get back his old
name when converting himself to a server thread. To not bother other
programs with unresolved references I do this with two function pointers
which are initialized in the main() programs for MR-AFS."