DELTA make-xdr-int64-support-consistent-now-that-its-everywhere-20010530 AUTHOR cg2v@andrew.cmu.edu
"an mr-afs related patch introduced xdr_afs_int64 into the code. This is a
good thing I suppose. The same patch also re-introduced the use of
xdr_long on some platforms, which isn't needed, and isn't really
appropriate. Said patch also only included the #defines for xdr_afs_int64
into rpc_main.c and not xdr.h. (it is my opinion that
rpc_main.c:write_int32_macros should only be used if kflag is set, but
transarc didn't do it that way, so I'm not going to change it)"