DELTA des-support-alpha-bsd-and-fix-test-programs-20020119 AUTHOR lha@e.kth.se
All BSDs are not I386s.

Also include a patch that make sure that data are alligned on alpha. All
bsd's will fix it up in kernel when used, but you'll get a huge performace
penalty

Use afs_int32 in the des programs to make sure they work and make them
build again.