DELTA eval-mountpoint-dont-return-eloop-20020819 AUTHOR zacheiss@mit.edu
Don't return ELOOP when crossing a mountpoint boundary from a BK volume to
a BK volume.  find treats it as a fatal error, and it's annoying not to be
able to naively use find to recurse through a volume with its backup volume
mounted.  Never noticed this before because we used to ignore this return
value.  Return ENODEV instead.