DELTA allow-fs-newcell-even-when-no-cells-exist-20020822 AUTHOR kolya@mit.edu
Make "fs newcell" work even when there are no pre-existing cells in
the kernel yet.  Previously, newcell tried to issue a GetCell pioctl
to figure out what version of the pioctl interface the kernel is
using.  Lacking such information (if the pioctl returns an error),
we now assume the newer interface.