DELTA rx-avoid-leaking-nwaiting-20030605 AUTHOR shadow@dementia.org FIXES
thanks to kolya@mit.edu for suggesting moving rather than adding the
decrement and confirming sanity.

basically if a call ended up call->error it was possible to queue_Remove()
it, drop the call, and not rx_nWaiting--; result should be observed as
rxdebug reporting waitprocs when threads were idle.