DELTA winnt-avoid-leak-and-memory-corruption-in-afscreds-20030915 AUTHOR jaltman@columbia.edu FIXES 2120
WINNT/talocale/TaLocale.h conditionally defines REALLOC() to
TaLocaleReallocFunction().

WINNT/client_creds/afscred.h  conditionally defines REALLOC() to
AfsCredsReallocFunction().

Unfortunately, AfsCredsReallocFunction is never used because REALLOC is
always defined first by an include of Talocale.h within afscreds.h.
This is a problem because it results in memory being allocated within a
DLL and then freed by the EXE.  This results in both a memory leak in
the DLL and memory corruption in the EXE.