DELTA jbuelhler-flexelint-aggregate-initialization-patches-20031129 AUTHOR jaltman@grand.central.org
This patch mainly makes explicit some initializations that were implicit.

There are several places where it looks like the missing initialization
may be a bug, and I have inserted comments to that effect in the
relevant patches.  Someone needs to look at them and supply
whatever is missing (if anything is).

In make_keyperm.c, an array was sized too large.
--
Joe Buehler