DELTA win32-name-event-objects-20040228 AUTHOR jaltman@mit.edu
This is a very bad leak of Event Objects.  Name all of our event objects
in an attempt to identify the source of the leak.  Apparently, the leak
is not coming from any Event Objects we are creating directly.

Did find a misallocation of event objects in the smb code for the
initial session object.  Fixed.