DELTA netbiosName-cellname-20040306 AUTHOR jaltman@mit.edu
Add support for a new form of automatic share name generation

        \\netbiosName\cellname

which means that on systems with the loopback adapter you can
now refer to the athena.mit.edu afs cell as \\afs\athena.mit.edu\

This is implemented within cm_FindShare().  If the share name is not
one of the special ones and cannot be found in the Submounts table,
check to see if it is a known cell name,  If so, create a path to
the share of /afs/sharename with the "sharename" being lower cased.
(I hope there are no cell names in mixed or upper case.  This won't
work.)