DELTA mutex-token-access-20040313 AUTHOR jaltman@mit.edu
This is an initial crack at one of the big problems with the existing
AFS client.  Multiple processes in a single session or in multiple
sessions (terminal server) can step on each others toes.  This is
because the dual use of RPC and SMB calls to perform operations.
Since nothing is done in a single synchronized step, there are plenty
of opportunities for data overwrites to occur.

The change was to add a Global Mutex within all of the ktc_XXXToken
calls.  The use of "Global\" prefixed kernel objects are only supported
on Win2000 and above.  This is fine because this is all we claim to
support.  But if someone was to decide to run this code on NT4 or Win95
anyway bad things will now happen.  This should probably be cleaned up
if someone has time.