DELTA cross-realm-obtain-tokens-afscreds-20040408 AUTHOR jaltman@mit.edu
If afscreds.exe you can now obtain credentials for cell "foo.com" with
credentials from "user@BAR.COM" when specifying a password.  This is a
first step since if there are already valid credentials for "user@BAR.COM"
the password should not be requested.  That would allow you to obtain
tokens for multiple cells with the same kerberos tgt.