DELTA reference-counts-20040511 AUTHOR jaltman@mit.edu
Correct VC reference counting.  (not that it is used for anything but
there is no reason the reference counting should be allowed to perpetually
increase)

Fix smb_FindFID.  Do not do anything if fid == 0 without SMB_CREATE_FLAG
being set.  Ensure that the event object name is unique by both the fid
and the vcid.