DELTA smb-notify-callback-expire-20040817 AUTHOR jaltman@mit.edu
When the callback expires on an scache item, if there is an smb client
listening for updates send it is a notification of change.  This will
force the smb client to refresh its listener which will in turn cause
a new callback to be registered.

Also, remove debugging statements from GetServer/PutServer