DELTA windows-maxmem-20041130 AUTHOR jaltman@mit.edu
There is currently a maximum cache size of 1.3GB.  The limit is imposed
by the largest contiguous block of unused memory within the 2GB process
space which can be assigned to the memory mapped file.  Unfortunately,
when the executable digital signature verification code is activated
Windows sees fit to further segment the process memory which in turn
reduces the size of the maximum cache file to less then 800MB.  If
larger cache sizes are desired, a new registry value should be set:

  HKLM\SOFTWARE\OpenAFS\Client (DWORD) "VerifyServiceSignature" = 0x0

Setting this value will disable the runtime verification of digital
signatures on afsd_service.exe and the afs dlls which it loads.  It
will not disable the the version number check on those same files.
The signature verification is not a security messure and is only meant
to enhance the ability to afsd_service.exe to detect potential
destablizing mixtures of DLLs from incompatible distributions.