DELTA STABLE14-windows-robust-freelance-20041126 AUTHOR jaltman@mit.edu
Implement new functions: cm_freelanceMountPointExists and
cm_freelanceSymlinkExists.  Use them along with other validity checks
in cm_freelanceAddMount and cm_freelanceAddSymlink to ensure that name
collisions do not occur and that empty strings are not valid file names.

A symlink may not have a name which would resolve to a valid cell name.
Doing so would prevent access to the cell.