DELTA aklog-segfault-fix-20050718 AUTHOR kekelley@iastate.edu FIXES 20198
I have built openafs 1.3.85 on an emt64 platform on RedHat Enterprise 4,
and found that the aklog supplied with 1.3.85 (not the one in the afs-krb5
conversion kit) consistently segfaults when aklog is called with the
"-path ...." parm. It works properly when called with no parms.

I traced this down to the call to the readlink() call in aklog_main.c,
when the params structure references were removed the nesting was
changed on an if statement that changed the logic and allowed the
do while loop to loop past the end and de-reference a NULL pointer.